Chunk 3: extract finance transactions into a service layer
Establish the service-layer pattern end-to-end on one route. New
services/transactionService.js holds all transaction business logic and
tenant data access; routes/finance/transactions.js (747 -> 179) becomes a
thin controller validating with Zod at the boundary and delegating to the
service, mirroring authService. Behaviour-preserving — existing finance
route tests pass unchanged. Document the convention in CLAUDE-STANDARDS and
record the larger follow-up routes (restore, members/crud) in KNOWN-ISSUES.
