+++ CHANGELOG.md
@@ -30,6 +30,17 @@ Format: `## [version] — YYYY-MM-DD` with bullet points per change.
     log PII/secrets in app logs.
 
 ### Changed
+- **Service-layer extraction for finance transactions (2026-06-14 review,
+  Chunk 3)** — introduced the marquee maintainability pattern on one route
+  end-to-end. New `backend/src/services/transactionService.js` now holds all
+  transaction business logic and data access (`list`/`get`/`create`/
+  `bulkSetPending`/`update`/`delete`/`refund`), and
+  `routes/finance/transactions.js` (747 → 179 lines) is a thin controller that
+  validates input with Zod at the route boundary and delegates to the service,
+  mirroring `services/authService.js`. Behaviour-preserving: the existing
+  finance route tests pass unchanged. The two larger offenders
+  (`routes/backup/restore.js`, `routes/members/crud.js`) are recorded in
+  `KNOWN-ISSUES.md` for the same treatment in follow-up sessions.
 - `CONTRIBUTING.md` Licensing section now points at the new `LICENSE`,
   `SECURITY.md`, and `docs/FromBeacon/README.md` instead of saying a licence is
   undecided.
